plugins/auth/requirements.txt
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
email-validator>=2.0.0,<3.0.0
redis>=4.5.0,<5.0.0